refactor(query): unify SDK planning path projection seam
